updates for chia-blockchain-1.8.1 (#8)
* updates for chia-blockchain-1.8.1

* pre-commit fixes

* fix variable names for fee and amt in cli

* use fee_str for all cmds

+++ tests/test_cb_manager.py
@@ -214,7 +214,7 @@ async def test_clawback(
 
         # Create a clawback with multiple xch coins
         spendable_balance = await wallet_maker.get_confirmed_balance()
-        coins = await wallet_maker.select_coins(spendable_balance)
+        coins = await wallet_maker.select_coins(uint64(spendable_balance))
         assert len(coins) > 1
         cb = await manager.create_cb_coin(spendable_balance, ph_taker, ph_maker, timelock, fee=0)
         res = await node_client.push_tx(cb)
